+++ svelte/Image.svelte
@@ -29,17 +29,20 @@
 		border-radius: inherit;
 	}
 	div.absolute {
+		width: 100%;
+		height: 100%;
 		position: absolute;
 		top: 0;
 		left: 0;
 	}
 	img {
+		width: 100%;
+		height: 100%;
+		object-fit: cover;
 		position: absolute;
 		top: 0;
 		left: 0;
-		width: 100%;
-		height: 100%;
 		border-radius: inherit;
-		object-fit: cover;
+		text-align: center;
 	}
 </style>