+++ svelte/index.js
@@ -1,3 +1,4 @@
+export { default as ButtonLink } from './ButtonLink.svelte';
 export { default as FeatherIcons } from './FeatherIcons.svelte';
 export { default as Icon } from './Icon.svelte';
 export { default as Image } from './Image.svelte';
@@ -11,3 +12,4 @@ export { default as ProgressBar } from './ProgressBar.svelte';
 export { default as ScrollTop } from './ScrollTop.svelte';
 export { default as SearchBar } from './SearchBar.svelte';
 export { default as ThemeSwitcher } from './ThemeSwitcher.svelte';
+export { default as WeavedImage } from './WeavedImage.svelte';